Move tool shed specific driver function to tool_shed.test
Should fix https://github.com/galaxyproject/galaxy/issues/18282
